Move new images in the root public assets
